remove old test_utils hack
