--- slosh/src/tests/utils.rs
@@ -1,1 +1,0 @@
-// This file is here to satisfy the compiler because of the path macro in the parent module.
